src/main/java/syu/autobiography/spring/audioupload/flask-app/requirements.txt
Flask==2.1.1
pydub==0.25.1
requests==2.26.0
SpeechRecognition==3.8.1
python-dotenv==0.19.2
Werkzeug==2.0.3
